weather-ai-app/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic-settings==2.10.1
httpx==0.28.1
alembic==1.16.5
python-dotenv==1.1.1
pytest==8.4.1
